fix queries

+++ src/eth/storage/postgres/queries/insert_account_batch.sql
@@ -1,6 +1,6 @@
 WITH account_updates
 AS (SELECT *
-    FROM UNNEST($1::bytea[], $2::bytea[], $3::numeric[], $4::numeric[], $5::int8[], $6::numeric[], $7::numeric[])
+    FROM UNNEST($1::bytea[], $2::bytea[], $3::numeric[], $4::numeric[], $5::numeric[], $6::numeric[], $7::numeric[])
     AS t(address, bytecode, new_balance, new_nonce, creation_block, original_balance, original_nonce)
 )
 INSERT INTO accounts (address, bytecode, latest_balance, latest_nonce, creation_block)
